backport changes made to chungking-core in stream-overlays
Fixes #146.
