+++ components/chungking-core/components/responsive-wrapper/index.ts
@@ -1,2 +1,4 @@
+export { default as ResponsiveIframe } from './ResponsiveIframe'
+export * from './ResponsiveIframe'
 export { default as ResponsiveWrapper } from './ResponsiveWrapper'
 export * from './ResponsiveWrapper'
