+++ lib/htmr-transform.tsx
@@ -4,7 +4,7 @@ import Image from 'next/image'
 import Link from 'next/link'
 import { HtmrOptions } from 'htmr/src/types'
 
-import { Anchor, Box, BoxProps, Iframe, MessageBox, ResponsiveWrapper } from '~/components/chungking-core'
+import { Anchor, Box, BoxProps, ResponsiveIframe, MessageBox, ResponsiveWrapper } from '~/components/chungking-core'
 import { H1, H2, H3, H4, H5, H6, P, UL, OL, LI, Blockquote, InlineCode, CodeBlock, Figure } from '~/modules/markdown'
 
 const htmrTransform: HtmrOptions['transform'] = {
@@ -74,7 +74,7 @@ const htmrTransform: HtmrOptions['transform'] = {
     const { title, mt: _mt, mb: _mb, allowFullScreen: _allowFullScreen, ...rest } = node
     return (
       <ResponsiveWrapper my="xl" mx={[null, null, null, null, -48]} borderRadius={6} boxShadow="single" overflow="hidden">
-        <Iframe title={title} allowFullScreen {...rest} />
+        <ResponsiveIframe title={title} allowFullScreen {...rest} />
       </ResponsiveWrapper>
     )
   },
